Move PluggableCommandLineOption to commoncode #2233
This way commoncode.cliutils has all the extended and enhanced CLI
Click utilities in one place.
