+++ src/cluecode/plugin_filter_clues.py
@@ -32,8 +32,8 @@ from itertools import chain
 import attr
 from six import string_types
 
-from plugincode import PluggableCommandLineOption
-from plugincode import POST_SCAN_GROUP
+from commoncode.cliutils import PluggableCommandLineOption
+from commoncode.cliutils import POST_SCAN_GROUP
 from plugincode.post_scan import PostScanPlugin
 from plugincode.post_scan import post_scan_impl
