+++ src/formattedcode/output_html.py
@@ -53,8 +53,8 @@ from commoncode.system import on_linux
 from commoncode.system import py2
 from commoncode.system import py3
 from formattedcode import FileOptionType
-from plugincode import PluggableCommandLineOption
-from plugincode import OUTPUT_GROUP
+from commoncode.cliutils import PluggableCommandLineOption
+from commoncode.cliutils import OUTPUT_GROUP
 from plugincode.output import output_impl
 from plugincode.output import OutputPlugin
