+++ src/summarycode/generated.py
@@ -35,8 +35,8 @@ from commoncode.datautils import Boolean
 from commoncode.text import toascii
 from plugincode.scan import ScanPlugin
 from plugincode.scan import scan_impl
-from plugincode import PluggableCommandLineOption
-from plugincode import OTHER_SCAN_GROUP
+from commoncode.cliutils import PluggableCommandLineOption
+from commoncode.cliutils import OTHER_SCAN_GROUP
 import typecode.contenttype
 
 """
