+++ tests/commoncode/test_cliutils.py
@@ -38,7 +38,7 @@ from click.testing import CliRunner
 from commoncode.testcase import FileDrivenTesting
 from commoncode.cliutils import fixed_width_file_name
 from commoncode.cliutils import GroupedHelpCommand
-from plugincode import PluggableCommandLineOption
+from commoncode.cliutils import PluggableCommandLineOption
 
 
 class TestUtils(FileDrivenTesting):
@@ -138,7 +138,7 @@ class TestGroupedHelpCommand(FileDrivenTesting):
 
         runner = CliRunner()
         result = runner.invoke(scan, ['--help'])
-        from plugincode import MISC_GROUP
+        from commoncode.cliutils import MISC_GROUP
         assert MISC_GROUP in result.output
         assert  '--opt   Help text for option' in result.output
 
@@ -152,11 +152,11 @@ class TestGroupedHelpCommand(FileDrivenTesting):
 
         runner = CliRunner()
         result = runner.invoke(scan, ['--help'])
-        from plugincode import MISC_GROUP
+        from commoncode.cliutils import MISC_GROUP
         assert MISC_GROUP + ':\n    --opt   Help text for option\n' in result.output
 
     def test_GroupedHelpCommand_help_with_group(self):
-        from plugincode import CORE_GROUP
+        from commoncode.cliutils import CORE_GROUP
 
         @click.command(name='scan', cls=GroupedHelpCommand)
         @click.option('--opt', is_flag=True, help='Help text for option',
